# Adding flask-caching to reduce load times