### What's Changed

Adds `flask-caching` to provide caching of requests for 1000 seconds (approx 15+ minutes)

| Before  | After   | 
|---|---|
| <img width="1680" alt="screen shot 2018-08-21 at 19 46 34" src="https://user-images.githubusercontent.com/1443700/44422931-8888e200-a57c-11e8-80f8-1b970f26556c.png"> | <img width="1680" alt="screen shot 2018-08-21 at 19 47 29" src="https://user-images.githubusercontent.com/1443700/44422957-9cccdf00-a57c-11e8-8f92-be47c6476535.png"> |

### Technical Description

Events do not frequently change so caching the results for 15 minutes can help reduce load times on all subsequent requests. In a local environment, this sees a reduction of 1.21s to 7ms. This doesn't account for network latency but it should help provide a small reduction in load times on the staging/production server.

This PR is a precursor to providing a static file on S3 containing the latest events.